Végpontok rendszerezése és Admin szerkesztése

+++ GyumiZoli_Backend/app/Http/Controllers/AuthController.php
@@ -19,9 +19,9 @@ class AuthController extends ResponseController
 
     public function setAdmin(Request $request){
 
-        // if (!Gate::allows("admin")) {
-        //     return response()->json(["error" => "Authentikációs hiba!", "message" => "Nincs jogosultság"]);
-        // }
+        if (!Gate::allows("admin")) {
+            return response()->json(["error" => "Authentikációs hiba!", "message" => "Nincs jogosultság"]);
+        }
 
         $user = User::find($request["id"]);
         $user->admin = $request["admin"];
@@ -33,9 +33,9 @@ class AuthController extends ResponseController
 
     public function updateUsers(Request $request)
     {
-        // if (!Gate::allows("admin")) {
-        //     return response()->json(["error" => "Authentikációs hiba!", "message" => "Nincs jogosultság"]);
-        // }
+        if (!Gate::allows("admin")) {
+            return response()->json(["error" => "Authentikációs hiba!", "message" => "Nincs jogosultság"]);
+        }
         $user = User::find($request["id"]);
         $user->name = $request["name"];
         $user->email = $request["email"];
@@ -48,9 +48,9 @@ class AuthController extends ResponseController
 
     public function destroyUser(Request $request)
     {
-        // if (!Gate::allows("admin")) {
-        //     return response()->json(["error" => "Authentikációs hiba!", "message" => "Nincs jogosultság"]);
-        // }
+        if (!Gate::allows("admin")) {
+            return response()->json(["error" => "Authentikációs hiba!", "message" => "Nincs jogosultság"]);
+        }
         $user = User::find($request["id"]);
         $user->delete();
         return response()->json($user); 
@@ -59,9 +59,9 @@ class AuthController extends ResponseController
 
     public function addUserAdmin(Request $request){
 
-        // if (!Gate::allows("admin")) {
-        //     return response()->json(["error" => "Authentikációs hiba!", "message" => "Nincs jogosultság"]);
-        // }
+        if (!Gate::allows("admin")) {
+            return response()->json(["error" => "Authentikációs hiba!", "message" => "Nincs jogosultság"]);
+        }
 
         $user = User::create([
             "name" => $request["name"],
@@ -76,44 +76,6 @@ class AuthController extends ResponseController
 
         return response()->json($user);
     }
-    public function updateUserAdmin(Request $request)
-    {
-        // if (!Gate::allows("admin")) {
-        //     return $this->sendError("Authentikációs hiba!","Nincs jogosultság",401);
-        // }
-
-        $user = User::findOrFail($request->input('id'));
-        $oldImage = $user->profile_picture;
-
-        if ($request->has('delete_image') && $request->input('delete_image') == true) {
-            if ($oldImage) {
-                $imagePath = str_replace(url('/storage/'), '', $oldImage);
-                if (Storage::disk('public')->exists($imagePath)) {
-                    Storage::disk('public')->delete($imagePath);
-                }
-            }
-            $user->profile_picture = null;
-        }
-
-        if ($request->hasFile('profile_picture')) {
-            if ($oldImage) {
-                $imagePath = str_replace(url('/storage/'), '', $oldImage);
-                if (Storage::disk('public')->exists($imagePath)) {
-                    Storage::disk('public')->delete($imagePath);
-                }
-            }
-            $user->profile_picture = $request->file('profile_picture')->store('profile_pictures', 'public');
-        }
-
-        $user->name = $request->input('name');
-        $user->email = $request->input('email');
-        $user->birth_date = $request->input('birth_date');
-        $user->address = $request->input('address');
-        $user->phone = $request->input('phone');
-        $user->admin = $request->input('admin');
-
-        $user->update();
-        return response()->json("Sikeres frissítés!");
-    }
+    
     
 }
